refactor: docstrings, URLs, list concatenation, file naming
