+++ cms/urls.py
@@ -15,19 +15,8 @@ The pattern(s) defined here serve the same Wagtail view that the library-defined
 from django.urls import re_path
 from wagtail import views
 from wagtail.coreutils import WAGTAIL_APPEND_SLASH
-from wagtail.api.v2.router import WagtailAPIRouter
 
 from cms.constants import COURSE_INDEX_SLUG, PROGRAM_INDEX_SLUG, WEBINAR_INDEX_SLUG
-from cms.wagtail_api import (
-    CustomPagesAPIViewSet,
-    CustomImagesAPIViewSet,
-    CustomDocumentsAPIViewSet,
-)
-
-api_router = WagtailAPIRouter("wagtailapi")
-api_router.register_endpoint("pages", CustomPagesAPIViewSet)
-api_router.register_endpoint("images", CustomImagesAPIViewSet)
-api_router.register_endpoint("documents", CustomDocumentsAPIViewSet)
 
 index_page_pattern = (
     rf"(?:{COURSE_INDEX_SLUG}|{PROGRAM_INDEX_SLUG}|{WEBINAR_INDEX_SLUG})"
