+++ cms/views.py
@@ -1,5 +1,5 @@
 """
-Customize the Wagtail API viewsets to restrict access to admin users
+Customize the Wagtail API viewsets to allow public access
 and to add additional filtering and metadata fields.
 """
 
@@ -9,12 +9,12 @@ from wagtail.api.v2.views import PagesAPIViewSet
 from wagtail.images.api.v2.views import ImagesAPIViewSet
 from wagtail.documents.api.v2.views import DocumentsAPIViewSet
 
-from .wagtail_api_filters import ReadableIDFilter
+from .filters import ReadableIDFilter
 
 
 class AllowAnyViewSetMixin:
     """
-    Mixin to override default permission to be authenticated.
+    Mixin to override default permission classes to allow public access.
     """
 
     permission_classes = (AllowAny,)
@@ -26,9 +26,10 @@ class CustomPagesAPIViewSet(AllowAnyViewSetMixin, PagesAPIViewSet):
     additional filtering and metadata fields.
     """
 
-    filter_backends = [ReadableIDFilter] + PagesAPIViewSet.filter_backends
-    meta_fields = PagesAPIViewSet.meta_fields + ["live", "last_published_at"]
-    listing_default_fields = PagesAPIViewSet.listing_default_fields + [
+    filter_backends = [ReadableIDFilter, *PagesAPIViewSet.filter_backends]
+    meta_fields = [*PagesAPIViewSet.meta_fields, "live", "last_published_at"]
+    listing_default_fields = [
+        *PagesAPIViewSet.listing_default_fields,
         "live",
         "last_published_at",
     ]
